Library/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
envparse==0.2.0
et-xmlfile==1.1.0
fastapi==0.109.0
fastapi-filter==1.1.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humps==0.2.2
idna==3.6
Mako==1.3.1
MarkupSafe==2.1.4
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
